Change the path to / and trigger deleteAccount in the index component
